webapi/webapi-messaging-w3c-tests/messaging/MmsMessage.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: MmsMessage</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#mmsmessage-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var filter = {
  type: "mms",
};
var filterOption = {
  sortBy: "date",
  sortBy: "descending",
  limit: 3
};

var t1 = async_test("Check if the messageID of MmsMessage is the identifier of the message", {timeout: 2000});

var t2 = async_test("Check if the type of MmsMessage is the type of message", {timeout: 2000});

var t3 = async_test("Check if the serviceID of MmsMessage is the messaging service id used to send / receive this message", {timeout: 2000});

var t4 = async_test("Check if the from of MmsMessage is the sender of the message", {timeout: 2000});

var t5 = async_test("Check if the timestamp of MmsMessage is the time the message reached the Multimedia Message Center", {timeout: 2000});

var t6 = async_test("Check if the expiry of MmsMessage is the number of seconds the message will be stored in the Multimedia Messaging Service Center and thus available for download", {timeout: 2000});

var t7 = async_test("Check if the read of MmsMessage is 'true' if the message has been marked as read", {timeout: 2000});

var t8 = async_test("Check if the read of MmsMessage is 'false' if the message has not been marked as read", {timeout: 2000});

var t9 = async_test("Check if the to of MmsMessage is an array containing the recipient(s) included in the 'To' field of the MMS message", {timeout: 2000});

var t10 = async_test("Check if the cc of MmsMessage is an array containing the recipient(s) included in the 'Cc' field of the MMS message", {timeout: 2000});

var t11 = async_test("Check if the bcc of MmsMessage is an array containing the recipient(s) included in the 'Bcc' field of the MMS message", {timeout: 2000});

var t12 = async_test("Check if the subject of MmsMessage is the subject of the MMS message", {timeout: 2000});

var t13 = async_test("Check if the smil of MmsMessage is the SMIL", {timeout: 2000});

var t14 = async_test("Check if the attachments of MmsMessage is the set of attachments of the MMS message", {timeout: 2000});

var t15 = async_test("Check if the state of MmsMessage is in the states of MmsState", {timeout: 2000});

var t16 = async_test("Check if the deliveryInfo of MmsMessage is an array with each of the items indicating the delivery status", {timeout: 2000});

setup(
  function () {
    message.mms.send({subject: "time", to: "5556"});
  }
)

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t1.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].messageID, "The messageID of MmsMessage");
      });
      t1.done();
    },
    function (error) {
      t1.step(function () {
        assert_unreached(error.message);
      });
      t1.done();
    }
  );
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t2.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].type, "mms", "The type of MmsMessage");
      });
      t2.done();
    },
    function (error) {
      t2.step(function () {
        assert_unreached(error.message);
      });
      t2.done();
    }
  );
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t3.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].serviceID, "sim0", "The serviceID of MmsMessage");
      });
      t3.done();
    },
    function (error) {
      t3.step(function () {
        assert_unreached(error.message);
      });
      t3.done();
    }
  );
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t4.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].from, "The from of MmsMessage");
      });
      t4.done();
    },
    function (error) {
      t4.step(function () {
        assert_unreached(error.message);
      });
      t4.done();
    }
  );
});

t5.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t5.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].timestamp, "The timestamp of MmsMessage");
      });
      t5.done();
    },
    function (error) {
      t5.step(function () {
        assert_unreached(error.message);
      });
      t5.done();
    }
  );
});

t6.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t6.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].expiry, "The expiry of MmsMessage");
      });
      t6.done();
    },
    function (error) {
      t6.step(function () {
        assert_unreached(error.message);
      });
      t6.done();
    }
  );
});

t7.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t7.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(typeof(msg.element[0].read), "boolean", "MmsMessage.read");
      });
      t7.done();
    },
    function (error) {
      t7.step(function () {
        assert_unreached(error.message);
      });
      t7.done();
    }
  );
});

t8.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t8.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_false(msg.element[0].read, "MmsMessage.read");
      });
      t8.done();
    },
    function (error) {
      t8.step(function () {
        assert_unreached(error.message);
      });
      t8.done();
    }
  );
});

t9.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t9.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].to, "The to of MmsMessage");
      });
      t9.done();
    },
    function (error) {
      t9.step(function () {
        assert_unreached(error.message);
      });
      t9.done();
    }
  );
});

t10.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t10.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(!!msg.element[0].cc, "The cc of MmsMessage");
      });
      t10.done();
    },
    function (error) {
      t10.step(function () {
        assert_unreached(error.message);
      });
      t10.done();
    }
  );
});

t11.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t11.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].bcc, "The bcc of MmsMessage");
      });
      t11.done();
    },
    function (error) {
      t11.step(function () {
        assert_unreached(error.message);
      });
      t11.done();
    }
  );
});

t12.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t12.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].subject, "time", "The subject of MmsMessage");
      });
      t12.done();
    },
    function (error) {
      t12.step(function () {
        assert_unreached(error.message);
      });
      t12.done();
    }
  );
});

t13.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t13.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(!!msg.element[0].smil, "The smil of MmsMessage");
      });
      t13.done();
    },
    function (error) {
      t13.step(function () {
        assert_unreached(error.message);
      });
      t13.done();
    }
  );
});

t14.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t14.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(msg.element[0].attachments, "The attachments of MmsMessage");
      });
      t14.done();
    },
    function (error) {
      t14.step(function () {
        assert_unreached(error.message);
      });
      t14.done();
    }
  );
});

t15.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t15.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_in_array(msg.element[0].state, ["not-downloaded", "fetching", "received", "draft", "sending", "sent", "failed"], "The state of MmsMessage");
      });
      t15.done();
    },
    function (error) {
      t15.step(function () {
        assert_unreached(error.message);
      });
      t15.done();
    }
  );
});

t16.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t16.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(msg.element[0].deliveryInfo, "The deliveryInfo of MmsMessage");
      });
      t16.done();
    },
    function (error) {
      t16.step(function () {
        assert_unreached(error.message);
      });
      t16.done();
    }
  );
});

</script>
